Feed functionality
Feed works! Settings and settings view were changed to accomodate. Max entries can be changed in settings. The only way to add feeds is by adjusting the feed.txt file in My Documents/MediaDrip/config/feed.txt.

Next I need to clean up code, write documentation and merge to master.  Afterwards, I'll need to update the view service to properly route so I can implement a media player on the Browse view.

The idea is to click on an item in your feed, it'll send the link address to the Browse view where the user can view the media.